MUC Cleaner - Refactored index to become index_renderer.
